Refactor IOMMU CSR code
Instead of having a separate IommuCsr class, we now hold the state of
the CSRs in member variables whose type corresponds to the specific CSR.
For example, the ddtp state is now held in ddtp_ member which is of type
Ddtp.

Each CSR has its own pair of methods for reading and writing. For
example, ddtp has readDdtp() and writeDdtp(). The readCsr() and
writeCsr() methods use the offset to invoke the appropriate method.