Make configFile option a comma-separated list
This allows for specifying multiple config files on the command line
which are processed from left to right using the merge_patch method.
This allows for overriding specific key/value pairs in the base config
file without having to modify it or create modified copies.

For example, if you want to change the mask for the misa CSR, you can
use the following file: { "csr": { "misa": { "mask": "0x201a8" } } }
And run: whisper --configfile whisper.json,overrides.json test.elf

Had to make a minor change to Aplic domain config parsing for this to
work: the value of null effectively deletes a key/vair pair when using
merge_patch, so "parent" has to become optional (omitting it defaults to
null, i.e. makes it the root domain).